requirements.txt
Pillow==6.0.0
PyPDF2==1.26.0
reportlab==3.5.21
